requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
blas=1.0=mkl
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py312h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py312h06a4308_0
fire=0.6.0=pyhd8ed1ab_0
freetype=2.12.1=h4a9f257_0
ftfy=6.1.3=pyhd8ed1ab_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.7=py312h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py312h06a4308_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libexpat=2.6.2=h59595ed_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=13.2.0=h77fa898_7
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=h5eee18b_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py312h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_1
mkl_fft=1.3.8=py312h5eee18b_0
mkl_random=1.2.4=py312hdb19cb5_0
mpmath=1.3.0=py312h06a4308_0
ncurses=6.5=h59595ed_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py312h06a4308_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.3.0=h4ab18f5_3
pillow=10.3.0=py312h5eee18b_0
pip=24.0=py312h06a4308_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.3=hab00c5b_0_cpython
python_abi=3.12=4_cp312
pytorch=2.3.0=py3.12_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py312h5eee18b_0
readline=8.2=h5eee18b_0
regex=2024.5.15=py312h9a8786e_0
requests=2.32.2=py312h06a4308_0
setuptools=69.5.1=py312h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py312h06a4308_0
tbb=2021.8.0=hdb19cb5_0
termcolor=2.4.0=pyhd8ed1ab_0
tk=8.6.14=h39e8969_0
torchaudio=2.3.0=py312_cu118
torchvision=0.18.0=py312_cu118
tqdm=4.66.4=py312he106c6f_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=py312h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.43.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.5=hc292b87_2
